feat(tensor): add ResetGradContext to tensor interface (#15)
feat(tensor): add ResetGradContext to tensor interface

- Feat!: add ResetGradContext to tensor interface.
In a model, each tensor requires to change its GradContext behavior in different batches/epochs. For instance, in prediction mode, no tensor should track the gradients. While in train mode, most of the tensors track gradients but some might not because of pre-training. Also, for the sake of fine-tuning and nature of models like GANs, tensors must be able to switch between these behaviors whenever needed.
- Refactor: replace forbidden with more meaningfull bpdirty.
bpdirty stands for back-propagation dirty. In back-propagation, a tensor can be visited many times leading to gradient accumulation. The first time it's visited, it would be marked as dirty. Thereafter, any calculation that uses it would not track gradients anymore.
- Fix: now every tensor must have GradContext.
The behavior of gradient tracking is now defined by two boolean properties: tracked and bpdirty. This is the reason that methods like Eq, which are not differentiable by nature, now initialize GradContext of their result.
- Refactor: move forbidden/required logic into gradtrack package.