fix(component): remove excessive interface nil checks (#24)
Fix!: nil value checks for interface types in component tree is removed.
this case mostly applies to tensor.Tensor and layers.Initializer types.
thus, for invalid state of an interface, behavior is up to the developer and panic might occure.
in this project, polymorphic behaviors are validated via contracts and the state of the polymorphic type is not validated. for cases like nil value, it must be exceptional like error interface where nil value has meaning in the business logic.
Refactor: BCE and MSE dimension validation error messages improved.