feat(tensor): add argmin and argmax reducers (#27)
- Feat!: add `Argmin` and `Argmax` to tensor interface.
these functions are not monitored by `gradtrack` package as they are discrete and non-differentiable.
- Refactor: execute *reducer* operations `reducerPair` type.
this type holds both `index` and `value` for each tensor element. when the final result is calculated on the whole tensor, it will be ***unwrapped*** into a single `float64` value based on the requirement; like the case for `max` and `min` methods. it should be noted that `index` is virtually a counter for tensor elements when we iterate them *linearly*, and it has nothing to do with their physical position.
- Fix: make `max` and `min` methods stable by considering equality case.