feat(optimizers): add momentum to SGD (#41)
Feat: add momentum configuration to SGD optimizer. a map[*tensor.Tensor]tensor.Tensor is used to track updates (velocities) of each optimized parameter. Go guarantees that pointers won't change as long as they are not candidates for garbage collection; therefore, it's safe to use them as the key type to map.
Fix: in order to repeat optimization in tests, ResetGradContext is invoked on updated tensor.