refactor(component): constant tensor generation (#66)
Refactor: use normal tensor initialization instead of Scale(0) or Pow(0).
In some components, a constant tensor must be used in the operation. Up until now, this was done by multiplying or powering a base tensor with 0 so the shape and device of that base tensor is retained. From now on, after being able to receive tensors' Device, this action can be optimized by using the normal tensor generation functions such as Full.